+++ pkg/mesh-networking/translation/decorators/decorators.go
