+++ pkg/mesh-networking/translation/istio/decorators/trafficpolicy/decorators.go
@@ -2,14 +2,14 @@ package trafficpolicy
 
 import (
 	"github.com/solo-io/service-mesh-hub/pkg/api/discovery.smh.solo.io/v1alpha1"
-	"github.com/solo-io/smh/pkg/mesh-networking/decorators"
+	"github.com/solo-io/smh/pkg/mesh-networking/translation/decorators"
 	"istio.io/api/networking/v1alpha3"
 )
 
 /*
 	Interface definitions for decorators which take TrafficPolicy as an input and
 	decorate a given output resource.
- */
+*/
 
 // TrafficPolicyDecorators modify the DestinationRule based on a TrafficPolicy which applies to the MeshService.
 type DestinationRuleDecorator interface {
