+++ pkg/mesh-networking/translation/istio/dependencies.go
@@ -2,7 +2,7 @@ package istio
 
 import (
 	skv1alpha1sets "github.com/solo-io/skv2/pkg/api/multicluster.solo.io/v1alpha1/sets"
-	"github.com/solo-io/smh/pkg/mesh-networking/decorators"
+	"github.com/solo-io/smh/pkg/mesh-networking/translation/decorators"
 	"github.com/solo-io/smh/pkg/mesh-networking/translation/istio/meshservice"
 	"github.com/solo-io/smh/pkg/mesh-networking/translation/utils/hostutils"
 )
