+++ pkg/mesh-networking/translation/istio/mesh/istio_mesh_translator.go
@@ -18,8 +18,8 @@ import (
 	v1 "github.com/solo-io/skv2/pkg/api/core.skv2.solo.io/v1"
 	"github.com/solo-io/skv2/pkg/ezkube"
 	"github.com/solo-io/smh/pkg/common/defaults"
-	"github.com/solo-io/smh/pkg/mesh-networking/decorators"
 	"github.com/solo-io/smh/pkg/mesh-networking/reporting"
+	"github.com/solo-io/smh/pkg/mesh-networking/translation/decorators"
 	"github.com/solo-io/smh/pkg/mesh-networking/translation/utils/hostutils"
 	"github.com/solo-io/smh/pkg/mesh-networking/translation/utils/metautils"
 	"github.com/solo-io/smh/pkg/mesh-networking/translation/utils/protoutils"
