+++ pkg/mesh-networking/translation/istio/meshservice/istio_mesh_service_translator.go
@@ -3,8 +3,8 @@ package meshservice
 import (
 	discoveryv1alpha1 "github.com/solo-io/service-mesh-hub/pkg/api/discovery.smh.solo.io/v1alpha1"
 	"github.com/solo-io/service-mesh-hub/pkg/api/networking.smh.solo.io/snapshot/input"
-	"github.com/solo-io/smh/pkg/mesh-networking/decorators"
 	"github.com/solo-io/smh/pkg/mesh-networking/reporting"
+	"github.com/solo-io/smh/pkg/mesh-networking/translation/decorators"
 	"github.com/solo-io/smh/pkg/mesh-networking/translation/istio/meshservice/destinationrule"
 	"github.com/solo-io/smh/pkg/mesh-networking/translation/istio/meshservice/virtualservice"
 	"github.com/solo-io/smh/pkg/mesh-networking/translation/utils/hostutils"
