+++ pkg/mesh-networking/translation/istio/meshservice/virtualservice/virtual_service_translator.go
@@ -2,7 +2,7 @@ package virtualservice
 
 import (
 	"github.com/solo-io/skv2/pkg/ezkube"
-	"github.com/solo-io/smh/pkg/mesh-networking/decorators"
+	"github.com/solo-io/smh/pkg/mesh-networking/translation/decorators"
 	"github.com/solo-io/smh/pkg/mesh-networking/translation/istio/decorators/trafficpolicy"
 
 	"reflect"
